requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-azure-ai>=0.1.0
azure-ai-inference>=1.0.0b3
azure-core>=1.30.0
pinecone-client>=3.0.0
langchain-pinecone>=0.1.0
pypdf>=4.2.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
cloudinary>=1.40.0
python-multipart>=0.0.9
slowapi>=0.1.9
